4.23

+++ part4/controllers/blogs.js
@@ -1,25 +1,37 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({}).populate('user', { username: 1, name: 1 })
+    // when we define schema with the ref
   response.json(blogs)
 })
 
+// login first, then can post blog with token
 blogsRouter.post('/', async (request, response) => {
-	const body = request.body
-
+  const body = request.body
+  const user = request.user // alr extracted from token via userExtractor
+  if (!user) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  
 	const blog = new Blog({
 		title: body.title,
 		author: body.author,
 		url: body.url,
 		likes: body.likes? body.likes : 0,
+    user: user.id
 	})
 
-	if(body.title === undefined || body.url === undefined){
-		response.status(400).end()
-	}else{
-		const savedBlog = await blog.save() 
+	if(!body.title || !body.url){
+		response.status(400).json({ error: 'title and url are required' })
+	} else{
+		const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save() 
 		
 		response.status(201).json(savedBlog)
 	}  
@@ -35,8 +47,22 @@ blogsRouter.get('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  const user = request.user // alr extracted from token via userExtractor
+  if (!user) {
+    return response.status(404).json({ error: 'token invalid' })
+  }
+
+  const blog = await Blog.findById(request.params.id) // to be deleted
+  if (!blog) {
+    return response.status(404).json({ error: 'blog not found' })
+  }
+
+  if (blog.user.toString() !== request.user.id.toString()) {
+    return response.status(401).json({ error: 'unauthorized to delete the blog'})
+  } else {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
